Jan.6th.2020.15th

+++ train/getmetalearner.py
@@ -73,9 +73,11 @@ def get_metalearner(filename):
                                                   dtype=torch.float64))
                         if hc2[n] is None:
                             hc2[n] = (torch.randn(1, meta_input.size(0), 2,
-                                                  device=device),
+                                                  device=device,
+                                                  dtype=torch.float64),
                                       torch.randn(1, meta_input.size(0), 2,
-                                                  device=device))
+                                                  device=device,
+                                                  dtype=torch.float64))
                         meta_output, hc1[n], hc2[n] = \
                             metalearner(meta_input, hc1[n], hc2[n])
